requirements.txt
pillow=2.4.0